update component folders
