--- client/src/components/ChatApp.js
@@ -1,20 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import Avatar from './Avatar';
-import InputBox from './InputBox';
-import MsgArea from './MsgArea';
-import Send from './Send';
-
-function ChatApp() {
-  return (
-    <>
-      <div className='message-window fixed-bottom'>
-        <MsgArea />
-        {/* <Avatar /> */}
-        <InputBox />
-        {/* <Send /> */}
-      </div>
-    </>
-  );
-}
-
-export default ChatApp;
